requirements.txt
asgiref==3.6.0
CacheControl==0.12.11
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-autoslug==1.9.8
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
lxml==4.9.2
msgpack==1.0.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
pywebcopy==7.0.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14

